refactor:  도메인 폴더의 controller, dto를 src/api/ 로 이동 및 경로 수정
