+++ src/api/slack-bot/dto/generate-report.dto.ts
