+++ src/auth/auth.module.ts
@@ -10,8 +10,8 @@ import { PassportModule } from '@nestjs/passport';
 import { JwtModule } from '@nestjs/jwt';
 import { ConfigModule } from '@nestjs/config';
 import { AuthService } from './auth.service';
-import { AuthController } from './auth.controller';
-import { CallbackController } from './callback.controller';
+import { AuthController } from '../api/auth/auth.controller';
+import { CallbackController } from '../api/auth/callback.controller';
 import { JwtStrategy } from './strategies/jwt.strategy';
 import { JwtAuthGuard, OptionalJwtAuthGuard } from './guards/jwt-auth.guard';
 import { UsersModule } from '../users/users.module';
