+++ src/health/health.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TerminusModule } from '@nestjs/terminus';
 import { MikroOrmModule } from '@mikro-orm/nestjs';
-import { HealthController } from './health.controller';
+import { HealthController } from '../api/health/health.controller';
 
 @Module({
   imports: [TerminusModule, MikroOrmModule],
